massive update, almost from scratch
The new web site aims at putting the focus on end-users rather than on research engineers.

Notably, the new site contains the tutorials for several container implementations, and
the glossary to explain the notriously confusing jargons such as "slirp".

The tutorials slightly overlap with the existing documentation in the upstream projects,
but the tutorials in this repo are (attempted to be) written in the same form across
several container implementations, for ease of comparison.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>